refactor code: split composables into several files
