+++ app/src/main/java/com/joerakhimov/todo/ui/task/SectionDivider.kt
@@ -1,0 +1,14 @@
+package com.joerakhimov.todo.ui.task
+
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun SectionDivider() {
+    HorizontalDivider(
+        thickness = 0.5.dp,
+        color = MaterialTheme.colorScheme.onSurface
+    )
+}
