+++ app/src/main/java/com/joerakhimov/todo/ui/tasks/AddTodoButton.kt
@@ -1,0 +1,26 @@
+package com.joerakhimov.todo.ui.tasks
+
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import com.joerakhimov.todo.R
+
+@Composable
+fun AddTodoButton(onClick: () -> Unit) {
+    FloatingActionButton(
+        onClick = onClick,
+        containerColor = MaterialTheme.colorScheme.primary,
+        shape = MaterialTheme.shapes.extraLarge,
+    ) {
+        Icon(
+            imageVector = Icons.Default.Add,
+            tint = Color.White,
+            contentDescription = stringResource(R.string.add_task),
+        )
+    }
+}
